requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1

# pip install djoser djangorestframework-simplejwt==4.7.2
# pip install webcolors